#269 - Pull request feedback
