+++ Python/Product/Debugger/Debugger/DebugEngine/AD7ProgramProvider.cs
@@ -14,6 +14,7 @@
 
 using System;
 using System.Runtime.InteropServices;
+using Microsoft.PythonTools.DkmDebugger;
 using Microsoft.VisualStudio;
 using Microsoft.VisualStudio.Debugger.Interop;
 
@@ -22,7 +23,7 @@ namespace Microsoft.PythonTools.Debugger.DebugEngine {
     // This registered interface allows the session debug manager (SDM) to obtain information about programs 
     // that have been "published" through the IDebugProgramPublisher2 interface.
     [ComVisible(true)]
-    [Guid(Guids.CLSID_ProgramProvider)]
+    [Guid(Guids.ProgramProviderId)]
     public class AD7ProgramProvider : IDebugProgramProvider2 {
         public AD7ProgramProvider() {
         }
